No more dEnv, get rid of bad sql interpolation
